requirements-deepagent.txt
vllm>=0.11.1
transformers
sentence-transformers
aiohttp
alfworld
beautifulsoup4
chardet
fastapi
fuzzywuzzy
func-timeout
lxml
nltk
numpy<2.0.0
openai
openai-whisper
pandas>=2.2.0
pdfplumber
PyPDF2
python-dateutil
python-docx
python-Levenshtein
python-pptx
pyyaml
requests
rouge
scipy>=1.13.0
spotipy
sympy
tqdm
deepagents
langchain_core
nest_asyncio
numba
